refactor: Remove usages of Persent Consent State
