+++ src/persistence.js
@@ -876,31 +876,6 @@ export default function _Persistence(mpInstance) {
             );
         }
     };
-
-    this.saveUserConsentStateToCookies = function(mpid, consentState) {
-        //it's currently not supported to set persistence
-        //for any MPID that's not the current one.
-        if (consentState || consentState === null) {
-            var persistence = self.getPersistence();
-            if (persistence) {
-                if (persistence[mpid]) {
-                    persistence[
-                        mpid
-                    ].con = mpInstance._Consent.ConsentSerialization.toMinifiedJsonObject(
-                        consentState
-                    );
-                } else {
-                    persistence[mpid] = {
-                        con: mpInstance._Consent.ConsentSerialization.toMinifiedJsonObject(
-                            consentState
-                        ),
-                    };
-                }
-                self.savePersistence(persistence);
-            }
-        }
-    };
-
     this.swapCurrentUser = function(
         previousMPID,
         currentMPID,
@@ -967,18 +942,6 @@ export default function _Persistence(mpInstance) {
             : this.getCookie();
 
         return persistence;
-    };
-
-    this.getConsentState = function(mpid) {
-        var persistence = self.getPersistence();
-
-        if (persistence && persistence[mpid] && persistence[mpid].con) {
-            return mpInstance._Consent.ConsentSerialization.fromMinifiedJsonObject(
-                persistence[mpid].con
-            );
-        } else {
-            return null;
-        }
     };
 
     this.getFirstSeenTime = function(mpid) {
